Move `RouteDescription` to its own module
