+++ servant-routes/src/Servant/API/Routes/Internal/Route.hs
@@ -37,34 +37,11 @@ import Lens.Micro.TH
 import Network.HTTP.Types.Method (Method)
 import "this" Servant.API.Routes.Auth
 import "this" Servant.API.Routes.Header
+import "this" Servant.API.Routes.Internal.Description
 import "this" Servant.API.Routes.Internal.Request
 import "this" Servant.API.Routes.Internal.Response
 import "this" Servant.API.Routes.Param
 import "this" Servant.API.Routes.Path
-
-{- | Description of a route. This will correspond to the Servant @Description@ combinator.
-
-It should  be noted that the 'HasRoutes' behaviour for @Description@ diverges from that in
-@servant-openapi3@, in the case that one EP has multiple @Description@ combinators.
-For example, given the following API:
-
-@
-type MyAPI =
-  "transaction" :> TransactionAPI
-    :\<|> "user" :> Description "User sub-API"
-          :> ( Description "Get my user" :> Get '[JSON] User
-              :\<|> "list" :> Get '[JSON] [User]
-             )
-@
-
-The @Operation@ that @servant-openapi@ generates for the @GET /user@ endpoint will have the two
-@Description@s 'mappend'-ed together: @"User sub-APIGet my user"@.
-
-The corresponding 'Route' will take the most specific 'RouteDescription': @"Get my user"@.
--}
-newtype RouteDescription = RouteDescription {unDescription :: T.Text}
-  deriving (Show)
-  deriving (Eq, IsString, Ord, Semigroup, Monoid, ToJSON, FromJSON) via T.Text
 
 {- | Short summary of a route. This will correspond to the Servant @Summary@ combinator.
 The behaviour described for 'RouteDescription' is the same for 'RouteSummary'.
